requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
pillow==9.1.1
    # via -r requirements.in
pyside6==6.3.1
    # via -r requirements.in
pyside6-addons==6.3.1
    # via pyside6
pyside6-essentials==6.3.1
    # via pyside6
shiboken6==6.3.1
    # via pyside6
